requirements/test.txt
-r common.txt

Flask-Testing==0.5.0

PyHamcrest==1.9.0
pytest==2.9.2

selenium==2.53.6
splinter==0.7.3
